tools/python/requirements.txt
tensorflow-cpu==2.13.0
numpy>=1.23,<2.0
Pillow>=9.5,<11.0
